requirements.txt
py-cord==1.7.3
discord-py-slash-command==3.0.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
typing-extensions==3.10.0.2
attrs==21.2.0
multidict==5.1.0
yarl==1.6.3
chardet==4.0.0
idna==3.2
configparser==5.2.0